+++ pan-file-service/src/main/java/top/quhailong/pan/file/controller/UploadFileController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
-import top.quhailong.pan.file.provider.UploadFileProvider;
+import top.quhailong.pan.file.service.IUploadFileService;
 import top.quhailong.pan.request.QuickUploadFileRequest;
 import top.quhailong.pan.request.UploadFileRequest;
 import top.quhailong.pan.utils.RestAPIResult;
@@ -30,7 +30,7 @@ public class UploadFileController {
     @Resource
     private HttpServletRequest httpServletRequest;
     @Autowired
-    private UploadFileProvider uploadFileProvider;
+    private IUploadFileService uploadFileService;
 
     /**
      * 上传文件
@@ -45,7 +45,7 @@ public class UploadFileController {
         StopWatch stopWatch = new StopWatch();
         stopWatch.start();
         logger.info("上传文件数据处理开始,request:{}", request);
-        RestAPIResult<String> result = uploadFileProvider.uploadFileHandle(request);
+        RestAPIResult<String> result = uploadFileService.uploadFileHandle(request);
         logger.info("上传文件数据处理结束,result:{}", result);
         stopWatch.stop();
         logger.info("上传文件调用时间,millies:{}", stopWatch.getTotalTimeMillis());
@@ -65,7 +65,7 @@ public class UploadFileController {
         StopWatch stopWatch = new StopWatch();
         stopWatch.start();
         logger.info("秒传文件数据处理开始,request:{}", request);
-        RestAPIResult<String> result = uploadFileProvider.quickUploadFileHandle(request);
+        RestAPIResult<String> result = uploadFileService.quickUploadFileHandle(request);
         logger.info("秒传文件数据处理结束,result:{}", result);
         stopWatch.stop();
         logger.info("秒传文件调用时间,millies:{}", stopWatch.getTotalTimeMillis());
@@ -84,7 +84,7 @@ public class UploadFileController {
         StopWatch stopWatch = new StopWatch();
         stopWatch.start();
         logger.info("上传文件(内部调用)数据处理开始");
-        RestAPIResult<String> result = uploadFileProvider.uploadHandle(file);
+        RestAPIResult<String> result = uploadFileService.uploadHandle(file);
         logger.info("上传文件(内部调用)数据处理结束,result:{}", result);
         stopWatch.stop();
         logger.info("上传文件调用时间,millies:{}", stopWatch.getTotalTimeMillis());
