+++ pan-file-service/src/main/java/top/quhailong/pan/file/service/impl/Md5ServiceImpl.java
@@ -1,28 +1,26 @@
-package top.quhailong.pan.file.provider;
+package top.quhailong.pan.file.service.impl;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.stereotype.Component;
-import top.quhailong.pan.file.service.FileService;
+import org.springframework.stereotype.Service;
+import top.quhailong.pan.file.dao.FileDao;
+import top.quhailong.pan.file.service.IMd5Service;
 import top.quhailong.pan.utils.RestAPIResult;
 
-/**
- * Md5数据处理类
- *
- * @author: quhailong
- * @date: 2019/9/25
- */
-@Component
-public class Md5Provider {
+import java.util.concurrent.TimeUnit;
+
+@Service
+public class Md5ServiceImpl implements IMd5Service {
     @Autowired
     private RedisTemplate<String, String> redisTemplate;
     @Autowired
-    private FileService fileService;
+    private FileDao fileDao;
 
+    @Override
     public RestAPIResult<String> md5CheckHandle(String fid, String md5) {
         RestAPIResult<String> panResult = new RestAPIResult<>();
-        redisTemplate.opsForValue().set("fileMd5:" + fid, md5, 259200);
-        Integer count = fileService.checkMd5Whether(md5);
+        redisTemplate.opsForValue().set("fileMd5:" + fid, md5, 259200, TimeUnit.SECONDS);
+        Integer count = fileDao.checkMd5Whether(md5);
         if (count > 0) {
             panResult.success(null);
         } else {
