+++ pan-regist-page/src/main/resources/static/js/forgetpass.js
@@ -111,8 +111,9 @@ var REGIST = {
 									new invokeSettime("#TANGRAM__PSP_3__verifyCodeSend");
 								} else {
 						$("#TANGRAM__PSP_3__verifyCodeSEError").html(data.respMsg);
+									change();
 								}
-							}
+							},
 						});
 		});
 		//密码处理
